Split arch-design section, add links on main page
The previously created architecture design has been stripped from any
concrete instructions, and now focuses solely on explaining the
architecture of the software. It is intended for contributors and those
who want to understand OpenIO SDS.

The newly created admin and operations guide is focused exclusively on the
practical applications, and is aimed at ops and those indending to
make OpenIO SDS work on their platform.

The links for both guides have also been added to the index page.